requirements.txt
africastalking==1.2.8
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
idna==3.10
python-decouple==3.8
requests==2.32.3
schema==0.7.7
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
